Weights for test functions
If not a boundary mesh, then the coefficients
for the test functions are identical to
those of the trial functions, i.e.
testWeights is polyWeights.

For boundary meshes, the testWeights
are not copies, but are modified by setting one row to be
zeros. This enforces that the test functions satisfy
the homogeneous dirichlet boundary conditions